Extract Reporter interface and declare CsvFileReporter
