+++ lib/reporting/Reporter.ts
@@ -1,49 +1,18 @@
 import {OptionValues} from 'commander';
-import {createWriteStream} from 'fs';
 import AnalyzedLine from '../analyzing/AnalyzedLine';
-import { logger } from '../tools/Logger';
 
-export default class Reporter {
+export default interface Reporter {
 
-  private readonly _options: OptionValues;
+  report(analyzedLines: AnalyzedLine[]): Promise<void>;
+}
 
-  output?: string;
+export abstract class AbstractReporter implements Reporter {
 
-  constructor(options: OptionValues) {
+  protected readonly _options: OptionValues;
+
+  protected constructor(options: OptionValues) {
     this._options = options;
-    this.output = options.output;
   }
 
-  async report(analyzedLines: AnalyzedLine[]) {
-    return new Promise((resolve, reject) => {
-      logger.info('--------------------------------------------------------------------------------');
-      logger.info('Phase: "Reporting"');
-      logger.info(`  - analyzed lines: ${analyzedLines.length}`);
-      logger.info();
-
-      const hrStart: [number, number] = process.hrtime();
-
-      analyzedLines.sort((a, b) => (a.index - b.index));
-
-      if (this.output) {
-        const outputStream = createWriteStream(this.output, {flags: 'w'});
-        analyzedLines.forEach((line, index) => {
-          outputStream.write(`${line.reference};${line.url};${line.status}`);
-          if (line.error) {
-            outputStream.write(`;${line.error};`);
-            line.comments.forEach((comment) => outputStream.write(`${comment}`));
-          }
-          if (index < analyzedLines.length - 1) {
-            outputStream.write('\n');
-          }
-        });
-        outputStream.end();
-        outputStream.on('finish', resolve);
-      }
-
-      const hrEnd: [number, number] = process.hrtime(hrStart);
-      logger.info(`Execution time (hr): ${hrEnd[0]}s ${hrEnd[1] / 1000000}ms`);
-      logger.info();
-    });
-  }
+  abstract report(analyzedLines: AnalyzedLine[]): Promise<void>;
 }
